+++ littletree/__init__.py
@@ -1,11 +1,13 @@
 from .basenode import BaseNode
 from .maxdepth import MaxDepth
 from .node import Node
+from .nodemixin import NodeMixin
 
 __all__ = [
     "BaseNode",
     "Node",
-    "MaxDepth"
+    "NodeMixin",
+    "MaxDepth",
 ]
 
-__version__ = "0.5.1"
+__version__ = "0.6.0"