+++ littletree/basenode.py
@@ -1,18 +1,20 @@
-from collections import namedtuple, deque
-from collections.abc import ValuesView
-from typing import Mapping, Iterable, Optional, Union, Any, Callable, Iterator, Generic, TypeVar, \
-    Hashable, Tuple, List
+from typing import Mapping, Iterable, Optional, Union, Any, Callable, Generic, TypeVar, \
+    Hashable, ValuesView, Tuple
 
 from .exceptions import DuplicateParentError, DuplicateChildError, LoopError
 from .nodepath import NodePath
+from .nodemixin import NodeMixin, NodeItem
 
 TNode = TypeVar("TNode", bound="BaseNode")
 TIdentifier = TypeVar("TIdentifier", bound=Hashable)
-NodePredicate = Callable[[TNode, "NodeItem"], bool]
 
 
-class BaseNode(Generic[TIdentifier]):
-    """This can serve as a base class but direct use is possible."""
+class BaseNode(Generic[TIdentifier], NodeMixin):
+    """Minimalistic node class that a user can inherit from.
+
+    Compared to Node this class is more primitive and less opinionated.
+    This class is not made abstract, so it's possible to use a BaseNode directly.
+    """
     __slots__ = "_identifier", "_parent", "_cdict", "_cvalues", "_path"
 
     # These can be changed in child classes
@@ -54,14 +56,6 @@ class BaseNode(Generic[TIdentifier]):
         return "".join(output)
 
     def __eq__(self: TNode, other: TNode):
-        if self is other:
-            return True
-        elif isinstance(other, self.__class__):
-            return self.identifier == other.identifier and self._cdict == other._cdict
-        else:
-            return NotImplemented
-
-    def __eq__(self: TNode, other: TNode):
         """Check if two trees are equal.
 
         Trees are equal if they have the same structure.
@@ -93,7 +87,7 @@ class BaseNode(Generic[TIdentifier]):
         :return:
         """
         old_node = self._cdict.get(new_identifier)
-        if not new_node.is_root:
+        if new_node._parent:
             if old_node is not new_node:
                 raise DuplicateParentError(new_node)
         else:
@@ -122,8 +116,8 @@ class BaseNode(Generic[TIdentifier]):
 
     @identifier.setter
     def identifier(self, new_identifier):
-        p = self.parent
-        old_identifier = self.identifier
+        p = self._parent
+        old_identifier = self._identifier
         if p is not None:
             if new_identifier in p._cdict:
                 raise DuplicateChildError(new_identifier, p)
@@ -138,9 +132,9 @@ class BaseNode(Generic[TIdentifier]):
 
     @parent.setter
     def parent(self, new_parent):
-        old_parent = self.parent
+        old_parent = self._parent
         if old_parent is not new_parent:
-            t = self.identifier
+            t = self._identifier
             if new_parent is None:
                 self._parent = None
             elif t in new_parent:
@@ -185,22 +179,6 @@ class BaseNode(Generic[TIdentifier]):
         except AttributeError:
             path = self._path = self.path_class(self)
             return path
-
-    @property
-    def root(self) -> TNode:
-        """Return root of tree."""
-        p, p2 = self, self._parent
-        while p2:
-            p, p2 = p2, p2._parent
-        return p
-
-    @property
-    def is_leaf(self) -> bool:
-        return not self._cdict
-
-    @property
-    def is_root(self) -> bool:
-        return self._parent is None
 
     def update(
         self,
@@ -253,7 +231,7 @@ class BaseNode(Generic[TIdentifier]):
             for node in other:
                 if node._parent is not None:
                     node = node.copy() if mode == "copy" else node.detach()
-                other_dict[node.identifier] = node
+                other_dict[node._identifier] = node
             if check_loop:
                 self._check_loop2(other_dict.values())
             self._update(other_dict)
@@ -267,7 +245,6 @@ class BaseNode(Generic[TIdentifier]):
             other.clear()
         else:
             raise TypeError("new_children should be mapping, iterable or other node")
-        return self
 
     def _update(self, other: Mapping[TIdentifier, TNode]):
         """Low-level update. Never use directly.
@@ -291,11 +268,11 @@ class BaseNode(Generic[TIdentifier]):
             node._parent = None
         return node
 
-    def copy(self, _memo=None) -> TNode:
+    def copy(self, _memo=None, keep=None) -> TNode:
         """Make a shallow copy or deepcopy if memo is passed."""
-        return self.transform(lambda _node, _item: BaseNode())
+        return self.transform(lambda _node, _item: BaseNode(), keep=keep)
 
-    def transform(self, f: Callable[[TNode, "NodeItem"], TNode], keep=None) -> TNode:
+    def transform(self, f: Callable[[TNode, NodeItem], TNode], keep=None) -> TNode:
         """Make a modified copy of a tree.
 
         :param f: How to modify each node
@@ -335,7 +312,6 @@ class BaseNode(Generic[TIdentifier]):
         for child in self.children:
             child._parent = None
         self._cdict.clear()
-        return self
 
     def sort_children(
         self,
@@ -353,7 +329,7 @@ class BaseNode(Generic[TIdentifier]):
         if key:
             nodes = sorted(self.children, key=key, reverse=reverse)
             self._cdict.clear()
-            self._cdict.update((n.identifier, n) for n in nodes)
+            self._cdict.update((n._identifier, n) for n in nodes)
         else:
             nodes = sorted(self._cdict.items(), reverse=reverse)
             self._cdict.clear()
@@ -361,144 +337,6 @@ class BaseNode(Generic[TIdentifier]):
         if recursive:
             for c in self.children:
                 c.sort_children(key=key, recursive=recursive, reverse=reverse)
-        return self
-
-    def iter_tree(
-        self,
-        keep: NodePredicate = None,
-        order: str = "pre",
-        with_item: bool = False,
-    ) -> Union[Iterator[TNode], Iterator[Tuple[TNode, "NodeItem"]]]:
-        """Iterate through all nodes of tree
-
-        :param keep: Predicate whether to continue iterating at node
-        :param order: Whether to iterate in pre/post or level-order
-        :param with_item: Whether to yield (node, node_item) instead of just the node
-        :return: All nodes.
-        """
-        item = NodeItem(0, 0)
-        if not keep or keep(self, item):
-            if order in ("pre", "level"):
-                yield (self, item) if with_item else self
-            yield from self.iter_descendants(keep, order=order, with_item=with_item)
-            if order == "post":
-                yield (self, item) if with_item else self
-
-    def iter_ancestors(self) -> Iterator[TNode]:
-        p = self._parent
-        while p:
-            yield p
-            p = p._parent
-
-    def iter_descendants(
-        self,
-        keep: NodePredicate = None,
-        order: str = "pre",
-        with_item: bool = False,
-    ) -> Union[Iterator[TNode], Iterator[Tuple[TNode, "NodeItem"]]]:
-        """Iterate through descendants
-
-        :param keep: Predicate whether to continue iterating at node
-        :param order: Whether to iterate in pre/post or level-order
-        :param with_item: Whether to yield (node, node_item) instead of just the node
-        :return: All descendants.
-        """
-        if order == "pre":
-            descendants = self._iter_descendants_pre(keep)
-        elif order == "post":
-            descendants = self._iter_descendants_post(keep)
-        elif order == "level":
-            descendants = self._iter_descendants_level(keep)
-        elif order == "post_rec":  # For checking if post returns correct result
-            descendants = self._iter_descendants_post_rec(keep)
-        else:
-            raise ValueError('order should be "pre", "post" or "level"')
-
-        if with_item:
-            return descendants
-        return (node for (node, with_item) in descendants)
-
-    def iter_siblings(self) -> Iterator[TNode]:
-        """Return siblings."""
-        if self._parent is not None:
-            return (child for child in self._parent.children if child != self)
-        else:
-            return iter(())
-
-    def iter_leaves(self) -> Iterator[TNode]:
-        """Yield leaf nodes."""
-        for node in self.iter_descendants():
-            if node.is_leaf:
-                yield node
-
-    def iter_levels(self) -> Iterator[List[TNode]]:
-        """Iterate levels.
-
-        The caller can modify level inplace and only the nodes remaining will be recursed on.
-        This can be used to prune the search.
-        Calling level.reverse() during the iteration step would get the levels in zigzag order.
-        """
-        level = [self]
-        while level:
-            yield level
-            level = [child for node in level for child in node.children]
-
-    def _iter_descendants_pre(self, keep):
-        nodes = deque((child, NodeItem(index, 1)) for (index, child) in enumerate(self.children))
-        while nodes:
-            node, item = nodes.popleft()
-            if not keep or keep(node, item):
-                yield node, item
-                next_nodes = [(child, NodeItem(index, item.depth + 1))
-                              for index, child in enumerate(node.children)]
-                nodes.extendleft(reversed(next_nodes))
-
-    def _iter_descendants_post_rec(self, keep, _depth=1):
-        # Same as _iter_descendants_post (kept as reference implementation)
-        for index, child in enumerate(self.children):
-            item = NodeItem(index, _depth)
-            if not keep or keep(child, item):
-                yield from child._iter_descendants_post_rec(keep=keep, _depth=_depth + 1)
-                yield child, item
-
-    def _iter_descendants_level(self, keep):
-        nodes = deque((child, NodeItem(index, 1)) for (index, child) in enumerate(self.children))
-        while nodes:
-            node, item = nodes.popleft()
-            if not keep or keep(node, item):
-                yield node, item
-                next_nodes = [(child, NodeItem(index, item.depth + 1))
-                              for index, child in enumerate(node.children)]
-                nodes.extend(next_nodes)
-
-    def _iter_descendants_post(self, keep, _depth=1):
-        nodes = iter([(child, NodeItem(index, 1))
-                      for (index, child) in enumerate(self.children)])
-        node, item = next(nodes, (None, None))
-        stack = []
-
-        while node or stack:
-            # Go down
-            keep_node = (keep is None or keep(node, item))
-            while keep_node and node.children:
-                children = iter([(child, NodeItem(index, item.depth + 1))
-                                 for (index, child) in enumerate(node.children)])
-                stack.append((item, nodes))
-                node, item = next(children)
-                nodes = children
-                keep_node = (keep is None or keep(node, item))
-            if keep_node:
-                yield node, item
-
-            # Go right or go up
-            parent = node._parent
-            node, item = next(nodes, (None, None))
-            while node is None and stack:
-                node = parent
-                item, nodes = stack.pop()
-                yield node, item
-                parent = node._parent
-                node, item = next(nodes, (None, None))
 
     def iter_together(self, other) -> Tuple[TNode, Optional[TNode]]:
         """Yield all nodes in self with similar nodes in other.
@@ -514,7 +352,7 @@ class BaseNode(Generic[TIdentifier]):
                 stack.append(other)
                 if other:
                     try:
-                        other = other[node.identifier]
+                        other = other[node._identifier]
                     except KeyError:
                         other = None
             yield node, other
@@ -543,6 +381,3 @@ class BaseNode(Generic[TIdentifier]):
         ancestor = next((child for child in others if id(child) in ancestors), None)
         if ancestor:
             raise LoopError(self, ancestor)
-
-
-NodeItem = namedtuple("NodeItem", ["index", "depth"])